moving tours into studentservices
