+++ hknweb/studentservices/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from hknweb.studentservices.models import Resume, ReviewSession
+from hknweb.studentservices.models import DepTour, Resume, ReviewSession
 
 
 @admin.register(Resume)
@@ -13,7 +13,6 @@ class ResumeAdmin(admin.ModelAdmin):
 
 @admin.register(ReviewSession)
 class ReviewSessionAdmin(admin.ModelAdmin):
-
     fields = [
         "name",
         "slug",
@@ -36,3 +35,27 @@ class ReviewSessionAdmin(admin.ModelAdmin):
     ]
     ordering = ["-created_at"]
     autocomplete_fields = ["created_by"]
+
+
+@admin.register(DepTour)
+class ToursAdmin(admin.ModelAdmin):
+    fields = [
+        "name",
+        "confirmed",
+        "datetime",
+        "email",
+        "phone",
+        "comments",
+        "deprel_comments",
+    ]
+    readonly_fields = ("name", "email", "phone", "comments")
+    list_display = (
+        "name",
+        "confirmed",
+        "email",
+        "datetime",
+        "date_submitted",
+        "phone",
+        "comments",
+        "deprel_comments",
+    )
