+++ hknweb/studentservices/urls.py
@@ -11,4 +11,5 @@ urlpatterns = [
     path("reviewsessions/<int:id>", views.reviewsession_details),
     path("reviewsessions/new", views.add_reviewsession),
     path("reviewsessions/<int:pk>/edit", views.ReviewSessionUpdateView.as_view(), name="reviewsession_edit"),
+    path("tours", views.tour, name="tours"),
 ]
