+++ hknweb/templates/base.html
@@ -105,7 +105,7 @@
       <span class="slash">/</span>
       <a href="">Course Guide</a>
       <span class="slash">/</span>
-      <a href="{% url 'tours:index' %}">Department Tours</a>
+      <a href="{% url 'studentservices:tours' %}">Department Tours</a>
       <span class="slash">/</span>
       <a href="{% url 'studentservices:resume' %}">Resume Critique</a>
     </div>
@@ -218,7 +218,7 @@
         <span class="slash">/</span>
         Course Guide
         <span class="slash">/</span>
-        <a href="{% url 'tours:index' %}">Department Tours</a>
+        <a href="{% url 'studentservices:tours' %}">Department Tours</a>
         <span class="slash">/</span>
         <a href="{% url 'studentservices:resume' %}">Resume Critique</a>
       </div>
