--- hknweb/tours/urls.py
@@ -1,7 +1,0 @@
-from django.urls import path
-from . import views
-
-app_name = "tours"
-urlpatterns = [
-    path("", views.tour, name="index"),
-]
