+++ hknweb/urls.py
@@ -38,7 +38,6 @@ urlpatterns = [
     path("studentservices/", include("hknweb.studentservices.urls")),
     path("serv/", include("hknweb.serv.urls")),
     path("tutoring/", include("hknweb.tutoring.urls")),
-    path("tours/", include("hknweb.tours.urls"), name="tours"),
     path("cand/", include("hknweb.candidate.urls")),
     path("pages/", include("hknweb.markdown_pages.urls")),
     path("markdownx/", include("markdownx.urls")),
